improve import states, apply consistency to imports. move to types folder
