+++ src/app/manifest.ts
@@ -7,8 +7,8 @@ export default function manifest(): MetadataRoute.Manifest {
     description: 'A Next.js PWA weather application.',
     start_url: '/',
     display: 'standalone',
-    background_color: '#ffffff',
-    theme_color: '#000000',
+    background_color: '#232229',
+    theme_color: '#232229',
     icons: [
       {
         src: '/icons/icon-192x192.png',
