+++ src/components/CitySearch/CitySearch.tsx
@@ -7,8 +7,8 @@ import React, { useState, useRef, useEffect } from 'react';
 
 import { AccuWeatherService } from '@/api/weather/accuweather.service';
 import { useDebounce } from '@/hooks/useDebounce';
-import { ICity } from '@/types/city.types';
-import './CitySearch.css';
+import { ICity } from '@/types/components/CitySearch.types';
+import '@/component/CitySearch/CitySearch.css';
 
 interface CitySearchProps {
   showLoading?: boolean;
