--- src/components/DailyEvolution/DailyEvolution.types.ts
@@ -1,19 +1,0 @@
-export interface DailyEvolutionData {
-  hour: number;
-  temperature: number;
-}
-
-export const mockDailyEvolutionData: DailyEvolutionData[] = [
-  { hour: 8, temperature: 3 },
-  { hour: 9, temperature: -2 },
-  { hour: 10, temperature: 4 },
-  { hour: 11, temperature: 11 },
-  { hour: 12, temperature: 6 },
-  { hour: 13, temperature: 2 },
-  { hour: 14, temperature: 8 },
-  { hour: 15, temperature: 24 },
-  { hour: 16, temperature: 6 },
-  { hour: 17, temperature: 2 },
-  { hour: 18, temperature: 8 },
-  { hour: 19, temperature: 24 },
-];
