+++ src/types/accuweather.types.ts
