+++ src/types/components/DailyEvolution.types.ts
@@ -1,0 +1,4 @@
+export interface DailyEvolutionData {
+  hour: number;
+  temperature: number;
+}
