+++ src/types/weatherAppTypes.ts
